Update README.md with detailed BiDOGE information
This update prioritizes information on the BiDOGE token, highlighting its support for large volumes of investors, aid for stray dogs, and recognition of the LGBTQ community. It also emphasizes its hyper-deflationary and AutoBOOST technology, cyber-security features, and involvement in charitable campaigns related to the COVID-19 pandemic. No team token or pre-sale is offered.